Fix: Исправления согласно рекомендациям.

+++ main/src/main/java/ru/practicum/ewm/locations/dto/LocationDto.java
@@ -9,6 +9,6 @@ import lombok.experimental.FieldDefaults;
 @NoArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class LocationDto {
-    Float lat;
-    Float lon;
+    Double lat;
+    Double lon;
 }
